Move util folder, move defaultBlocks to util
