+++ src/components/SlateEditor/plugins/bodybox/index.js
@@ -9,7 +9,9 @@
 import React from 'react';
 import { Block } from 'slate';
 import SlateBodyBox from './SlateBodyBox';
-import { defaultBlock } from '../../schema';
+import { defaultBlocks } from '../../utils';
+
+const { defaultBlock } = defaultBlocks;
 
 export const defaultBodyBoxBlock = () =>
   Block.create({
