+++ src/components/SlateEditor/plugins/detailsbox/index.js
@@ -9,7 +9,7 @@
 import React from 'react';
 import { Block } from 'slate';
 import DetailsBox from './DetailsBox';
-import { defaultBlock } from '../../schema';
+import { defaultBlocks } from '../../utils';
 
 const detailsBlock = [
   {
@@ -36,7 +36,7 @@ export default function createDetails() {
     if (node.type !== 'details') return null;
     if (!node.nodes.last().type) return null;
     if (!node.nodes.last().isVoid) return null;
-    const block = Block.create(defaultBlock);
+    const block = Block.create(defaultBlocks.defaultBlock);
     return change => {
       change.insertNodeByKey(node.key, node.nodes.size, block);
     };
