+++ src/components/SlateEditor/plugins/footnote/EditFootnote.jsx
@@ -12,7 +12,7 @@ import { injectT } from 'ndla-i18n';
 import Types from 'slate-prop-types';
 import { compose } from 'redux';
 import FootnoteForm, { getInitialModel } from './FootnoteForm';
-import connectLightbox from '../utils/connectLightbox';
+import connectLightbox from '../../utils/connectLightbox';
 import { TYPE } from '.';
 
 class EditFootnote extends Component {
