+++ src/components/SlateEditor/utils/checkSelectionForType.js
