+++ src/components/SlateEditor/utils/isNodeInCurrentSelection.js
