+++ src/components/SlateEditor/utils/isTypeinSelection.js
