+++ src/containers/SearchPage/components/results/SearchAudio.jsx
@@ -15,6 +15,7 @@ import { toEditAudio } from '../../../../util/routeHelpers';
 import { AudioResultShape } from '../../../../shapes';
 import { searchClasses } from '../../SearchContainer';
 
+// NOT USED, can be deleted?
 const SearchAudio = ({ audio, locale, t }) => (
   <div {...searchClasses('result')}>
     <div {...searchClasses('image')}>
